* release/cleanup:
  possibility to use clang_complete
  update sublime project file
  remove unnessecary variable
  cleanup localsearch
  remove of scatter search data type
  cleanup cmake
  added gurobi 702
  changed name + added print functions to wct.h
  removed print functions from main file
  added print.c
  added some changes

+++ src/greedy.c
@@ -76,7 +76,7 @@ static int solution_set_c(solution *sol) {
     sol->tw = 0;
     sol->b = 0;
 
-    for (unsigned i = 0; i < sol->nmachines; ++i) {
+    for (int i = 0; i < sol->nmachines; ++i) {
         sol->part[i].c = 0;
         sol->part[i].tw = 0;
         sol->part[i].key = i;
@@ -85,7 +85,7 @@ static int solution_set_c(solution *sol) {
         binomial_heap_insert(heap, sol->part + i);
     }
 
-    for (unsigned i = 0; i < sol->njobs; ++i) {
+    for (int i = 0; i < sol->njobs; ++i) {
         j = sol->perm[i];
         tmp = (partlist *)binomial_heap_pop(heap);
         j->index = tmp->machine->len;
@@ -111,7 +111,7 @@ CLEAN:
 int construct_spt(wctproblem *prob, solution *sol) {
     int val = 0;
 
-    for (unsigned i = 0; i < prob->njobs; ++i) {
+    for (int i = 0; i < prob->njobs; ++i) {
         sol->perm[i] = prob->ojobarray[i];
     }
 
@@ -127,7 +127,7 @@ CLEAN:
 int construct_edd(wctproblem *prob, solution *sol) {
     int val = 0;
 
-    for (unsigned i = 0; i < prob->njobs; ++i) {
+    for (int i = 0; i < prob->njobs; ++i) {
         sol->perm[i] = prob->ojobarray[i];
     }
 
@@ -142,7 +142,7 @@ CLEAN:
 int construct_random(wctproblem *prob, solution *sol, GRand *rand_uniform) {
     int val = 0;
 
-    for (unsigned i = 0; i < prob->njobs; ++i) {
+    for (int i = 0; i < prob->njobs; ++i) {
         sol->perm[i] = prob->ojobarray[i];
     }
 
@@ -265,7 +265,7 @@ static void perturb_swap(solution *         sol,
                          int                l2,
                          GRand *            rand_uniform) {
     int       m1, m2;
-    int       i1 = 0, i2 = 0;
+    unsigned       i1 = 0, i2 = 0;
     int       nmachines = sol->nmachines;
     Job **    tmp1 = (Job **)NULL;
     Job **    tmp2 = (Job **)NULL;
@@ -282,7 +282,7 @@ static void perturb_swap(solution *         sol,
     part1 = sol->part + m1;
     part2 = sol->part + m2;
 
-    if ((int)part1->machine->len <= l1 || part2->machine->len <= l2) {
+    if (part1->machine->len <= l1 || part2->machine->len <= l2) {
         return;
     }
 
@@ -299,7 +299,7 @@ static void perturb_swap(solution *         sol,
         assert(i1 < part1->machine->len - l1);
     }
 
-    for (unsigned i = 0; i < l1; ++i) {
+    for (int i = 0; i < l1; ++i) {
         tmp1[i] = (Job *)g_ptr_array_index(part1->machine, i1);
         g_ptr_array_remove_index(part1->machine, i1);
     }
@@ -309,7 +309,7 @@ static void perturb_swap(solution *         sol,
         assert(i2 < part2->machine->len - l2);
     }
 
-    for (unsigned i = 0; i < l2; ++i) {
+    for (int i = 0; i < l2; ++i) {
         tmp2[i] = (Job *)g_ptr_array_index(part2->machine, i2);
         g_ptr_array_remove_index(part2->machine, i2);
     }
@@ -318,11 +318,11 @@ static void perturb_swap(solution *         sol,
         i2 = g_random_int_range(0, part2->machine->len);
         assert(i2 < part2->machine->len);
 
-        for (unsigned i = 0; i < l1; ++i) {
+        for (int i = 0; i < l1; ++i) {
             g_ptr_array_insert(part2->machine, i2 + i, tmp1[i]);
         }
     } else {
-        for (unsigned i = 0; i < l1; ++i) {
+        for (int i = 0; i < l1; ++i) {
             g_ptr_array_add(part2->machine, tmp1[i]);
         }
     }
@@ -331,11 +331,11 @@ static void perturb_swap(solution *         sol,
         i1 = g_random_int_range(0, part1->machine->len);
         assert(i1 < part1->machine->len);
 
-        for (unsigned i = 0; i < l2; ++i) {
+        for (int i = 0; i < l2; ++i) {
             g_ptr_array_insert(part1->machine, i1 + i, tmp2[i]);
         }
     } else {
-        for (unsigned i = 0; i < l2; ++i) {
+        for (int i = 0; i < l2; ++i) {
             g_ptr_array_add(part1->machine, tmp2[i]);
         }
     }
@@ -365,29 +365,29 @@ void Perturb(solution *sol, local_search_data *data, GRand *rand_uniform) {
     int L;
     L = g_rand_int_range(rand_uniform, 0, 3);
 
-    for (unsigned i = 0; i < L; ++i) {
+    for (int i = 0; i < L; ++i) {
         perturb_swap(sol, data, 1, 2, rand_uniform);
     }
 
     L = g_rand_int_range(rand_uniform, 0, 3);
 
-    for (unsigned i = 0; i < L; ++i) {
+    for (int i = 0; i < L; ++i) {
         perturb_swap(sol, data, 1, 3, rand_uniform);
     }
 
     L = g_rand_int_range(rand_uniform, 0, 3);
 
-    for (unsigned i = 0; i < L; ++i) {
+    for (int i = 0; i < L; ++i) {
         perturb_swap(sol, data, 2, 2, rand_uniform);
     }
 
     L = g_rand_int_range(rand_uniform, 0, 3);
 
-    for (unsigned i = 0; i < L; ++i) {
+    for (int i = 0; i < L; ++i) {
         perturb_swap(sol, data, 2, 3, rand_uniform);
     }
 
-    for (unsigned i = 0; i < sol->nmachines; ++i) {
+    for (int i = 0; i < sol->nmachines; ++i) {
         sol->part[i].used = 1;
     }
 
@@ -421,7 +421,7 @@ int heuristic_rpup(wctproblem *prob) {
     CCcheck_NULL_2(prob->opt_sol, "Failed to allocate memory");
     solution_update(prob->opt_sol, sol);
 
-    for (unsigned i = 0; i < IR; ++i) {
+    for (int i = 0; i < IR; ++i) {
         sol1 = solution_alloc(prob->nmachines, prob->njobs, prob->off);
         CCcheck_NULL_2(sol1, "Failed to allocate memory");
         val = construct_random(prob, sol1, rand_uniform);
@@ -431,7 +431,7 @@ int heuristic_rpup(wctproblem *prob) {
         local_search_create_g(sol1, data_RS);
         solution_update(sol, sol1);
 
-        for (unsigned j = 0; j < ILS; ++j) {
+        for (int j = 0; j < ILS; ++j) {
             RVND(sol1, data_RS);
 
             if (sol1->tw < sol->tw) {
